requirements.txt
chainer==3.1.0
numpy==1.13.1
Pillow==4.2.1
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
tensorboard-chainer==0.2.8.1
tqdm==4.14.0
pytz==2017.2
scipy==0.19.1
